Added lquery helper/wrapper for getting info
lquery.py is a script that attempts to make gathering info from the
fireworks database a bit easier.

It wraps the lpad command with new abilities, namely the ability to
specify both the name and state of a firework without having to write
the MongoDB query.

E.g. lquery.py get_fws -s RUNNING -n FSD_CosmicRun.larnd

Use --help (or -h) for the documenation.